+++ src/paco/stack/stack.py
@@ -769,10 +769,23 @@ your cache may be out of sync. Try running again the with the --nocache option.
 
     def get_outputs_key_from_ref(self, ref):
         "Return a key for an output from a Reference object"
+        # Secrets .jsonfield ref special case
+        if schemas.ISecretsManagerSecret.providedBy(ref.resource):
+            if ref.type == 'netenv':
+                if len(ref.parts) == 10:
+                    new_ref = ref.parts[:-2]
+                    new_ref.append('arn')
+                    new_ref = '.'.join(new_ref)
+                    ref = references.Reference(f'paco.ref {new_ref}')
+            else:
+                raise NotImplemented('Only netenv ref types supported for Secrets')
+
         for stack_output_config in self.stack_output_config_list:
             if stack_output_config.config_ref == ref.ref:
                 return stack_output_config.key
+
         # raise an error if no key was found
+        breakpoint()
         message = self.get_stack_error_message()
         message += "Error: Unable to find outputs key for ref: {}\n".format(ref.raw)
         raise StackException(